requirements.txt
opencv-python~=4.5.3
pyserial~=3.5
